pyproject.toml
[project]
name = "telegram-sales-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.14",
    "ddgs>=9.12.1",
    "fastapi[standard]>=0.135.3",
    "mistralai>=2.3.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
]
